LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/pragma-directives/attr-meta-http-equiv-refresh/parsing-expected.txt


Harness Error (TIMEOUT), message = null

FAIL <meta>: "" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: ""
TIMEOUT <meta>: "1" Test timed out
PASS Refresh header: "1"
TIMEOUT <meta>: "1 " Test timed out
PASS Refresh header: "1 "
TIMEOUT <meta>: "1\t" Test timed out
PASS Refresh header: "1\t"
TIMEOUT <meta>: "1\r" Test timed out
TIMEOUT <meta>: "1\n" Test timed out
TIMEOUT <meta>: "1\f" Test timed out
TIMEOUT <meta>: "1;" Test timed out
PASS Refresh header: "1;"
TIMEOUT <meta>: "1," Test timed out
PASS Refresh header: "1,"
TIMEOUT <meta>: "1; url=foo" Test timed out
PASS Refresh header: "1; url=foo"
TIMEOUT <meta>: "1, url=foo" Test timed out
PASS Refresh header: "1, url=foo"
TIMEOUT <meta>: "1 url=foo" Test timed out
PASS Refresh header: "1 url=foo"
TIMEOUT <meta>: "1;\turl=foo" Test timed out
PASS Refresh header: "1;\turl=foo"
TIMEOUT <meta>: "1,\turl=foo" Test timed out
PASS Refresh header: "1,\turl=foo"
TIMEOUT <meta>: "1\turl=foo" Test timed out
PASS Refresh header: "1\turl=foo"
TIMEOUT <meta>: "1;\rurl=foo" Test timed out
TIMEOUT <meta>: "1,\rurl=foo" Test timed out
TIMEOUT <meta>: "1\rurl=foo" Test timed out
TIMEOUT <meta>: "1;\nurl=foo" Test timed out
TIMEOUT <meta>: "1,\nurl=foo" Test timed out
TIMEOUT <meta>: "1\nurl=foo" Test timed out
TIMEOUT <meta>: "1;\furl=foo" Test timed out
TIMEOUT <meta>: "1,\furl=foo" Test timed out
TIMEOUT <meta>: "1\furl=foo" Test timed out
FAIL <meta>: "1url=foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "1url=foo"
FAIL <meta>: "1x;url=foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "1x;url=foo"
TIMEOUT <meta>: "1 x;url=foo" Test timed out
PASS Refresh header: "1 x;url=foo"
TIMEOUT <meta>: "1;;url=foo" Test timed out
PASS Refresh header: "1;;url=foo"
TIMEOUT <meta>: "  1  ;  url  =  foo" Test timed out
PASS Refresh header: "  1  ;  url  =  foo"
TIMEOUT <meta>: "  1  ,  url  =  foo" Test timed out
PASS Refresh header: "  1  ,  url  =  foo"
TIMEOUT <meta>: "  1  ;  foo" Test timed out
PASS Refresh header: "  1  ;  foo"
TIMEOUT <meta>: "  1  ,  foo" Test timed out
PASS Refresh header: "  1  ,  foo"
TIMEOUT <meta>: "  1  url  =  foo" Test timed out
PASS Refresh header: "  1  url  =  foo"
TIMEOUT <meta>: "1; url=foo " Test timed out
PASS Refresh header: "1; url=foo "
TIMEOUT <meta>: "1; url=f\to\no" Test timed out
TIMEOUT <meta>: "1; url=\"foo\"bar" Test timed out
PASS Refresh header: "1; url=\"foo\"bar"
TIMEOUT <meta>: "1; url='foo'bar" Test timed out
PASS Refresh header: "1; url='foo'bar"
TIMEOUT <meta>: "1; url=\"foo'bar" Test timed out
PASS Refresh header: "1; url=\"foo'bar"
TIMEOUT <meta>: "1; url foo" Test timed out
PASS Refresh header: "1; url foo"
TIMEOUT <meta>: "1; urlfoo" Test timed out
PASS Refresh header: "1; urlfoo"
TIMEOUT <meta>: "1; urfoo" Test timed out
PASS Refresh header: "1; urfoo"
TIMEOUT <meta>: "1; ufoo" Test timed out
PASS Refresh header: "1; ufoo"
TIMEOUT <meta>: "1; \"foo\"bar" Test timed out
PASS Refresh header: "1; \"foo\"bar"
FAIL <meta>: "; foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "; foo"
FAIL <meta>: ";foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: ";foo"
FAIL <meta>: ", foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: ", foo"
FAIL <meta>: ",foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: ",foo"
FAIL <meta>: "foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "foo"
FAIL <meta>: "+1; url=foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "+1; url=foo"
FAIL <meta>: "-1; url=foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "-1; url=foo"
FAIL <meta>: "+0; url=foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "+0; url=foo"
FAIL <meta>: "-0; url=foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "-0; url=foo"
TIMEOUT <meta>: "0; url=foo" Test timed out
PASS Refresh header: "0; url=foo"
FAIL <meta>: "+1; foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "+1; foo"
FAIL <meta>: "-1; foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "-1; foo"
FAIL <meta>: "+0; foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "+0; foo"
FAIL <meta>: "-0; foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "-0; foo"
TIMEOUT <meta>: "0; foo" Test timed out
PASS Refresh header: "0; foo"
FAIL <meta>: "+1" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "+1"
FAIL <meta>: "-1" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "-1"
FAIL <meta>: "+0" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "+0"
FAIL <meta>: "-0" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "-0"
TIMEOUT <meta>: "0" Test timed out
PASS Refresh header: "0"
TIMEOUT <meta>: "1.9; url=foo" Test timed out
PASS Refresh header: "1.9; url=foo"
TIMEOUT <meta>: "1.9..5.; url=foo" Test timed out
PASS Refresh header: "1.9..5.; url=foo"
TIMEOUT <meta>: ".9; url=foo" Test timed out
PASS Refresh header: ".9; url=foo"
TIMEOUT <meta>: "0.9; url=foo" Test timed out
PASS Refresh header: "0.9; url=foo"
TIMEOUT <meta>: "0...9; url=foo" Test timed out
PASS Refresh header: "0...9; url=foo"
TIMEOUT <meta>: "0...; url=foo" Test timed out
PASS Refresh header: "0...; url=foo"
FAIL <meta>: "1e0; url=foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "1e0; url=foo"
FAIL <meta>: "1e1; url=foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "1e1; url=foo"
FAIL <meta>: "10e-1; url=foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "10e-1; url=foo"
FAIL <meta>: "-0.1; url=foo" assert_equals: expected "refresh.sub.html" but got "{\"error\": {\"message\": \"No such file or directory\", \"code\": 404}}"
PASS Refresh header: "-0.1; url=foo"
